requirements.txt
requests==2.32.5
python-dotenv==1.1.1
python-telegram-bot==13.15
redis==6.4.0